Added dates to the Tasks dashboard

+++ src/routes/tasks/+server.js
@@ -7,21 +7,23 @@ import getData from './getFire.js'
 
 
 
-export async function POST(requestEvent){
+export async function POST(requestEvent){ //create a task
     requestEvent.preventDefault()
     const formData = new FormData(requestEvent.target)
     const Usuario = formData.get('Usuario')
     const Nombre = formData.get('Nombre')
     const Descripcion = formData.get('Descripcion')
+    const Date = formData.get('date')
     const NewTask = {
         Usuario:Usuario,
         Nombre: Nombre,
-        Descripcion:Descripcion
+        Descripcion:Descripcion,
+        Date:Date
     }
     console.log(JSON.stringify(NewTask))
 
     const userRef = firestore.collection('Tasks').doc()
-    await userRef.set({ Usuario, Descripcion, Nombre });
+    await userRef.set({ Usuario, Descripcion, Nombre, Date });
 
     window.location.href = "/tasks"
 
@@ -31,20 +33,21 @@ export async function DELETE(requestEvent){
     requestEvent.preventDefault()
     const formData = new FormData(requestEvent.target)
     const item = formData.get('item').split(',')
+    console.log(item)
     let Nombre = item[0]
     let Desc = item[1]
     let User = item[2]
+    let Date = item[3]
     let Database = await getData(); 
 
 
     for(var i = 0; i < Database.length; i ++){
         let current  = Database[i]
-        console.log(current)
         let DB_Nombre = current.Nombre
         let DB_Descripcion = current.Descripcion
         let DB_Usuario = current.Usuario
-
-        if (DB_Nombre === Nombre && DB_Descripcion === Desc && DB_Usuario === User){
+        let DB_Date = current.Date
+        if (DB_Nombre === Nombre && DB_Descripcion === Desc && DB_Usuario === User && DB_Date == Date){
             console.log(current.id)
             let docRef = firestore.collection('Tasks').doc(current.id);
             docRef.delete().then(() => {
